requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
deta==1.1.0
fastapi==0.85.1
idna==3.4
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
